Refactor low level redcap details into helper class RedcapSessionManager - use helper class for ClinicalDataSource and MetadataManager - capture the token filter mode and check for unexpected changes
