+++ redcap/redcap_source/src/main/java/org/mskcc/cmo/ks/redcap/source/internal/ClinicalDataSourceRedcapImpl.java
@@ -23,11 +23,11 @@
  *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU Affero General Public License for more details.
  *
  * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
 */
 package org.mskcc.cmo.ks.redcap.source.internal;
 
@@ -36,17 +36,12 @@ import com.fasterxml.jackson.databind.node.ObjectNode;
 import java.io.*;
 import java.net.*;
 import java.util.*;
-import java.util.function.Predicate;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 import org.apache.commons.text.StringEscapeUtils;
 import org.apache.log4j.Logger;
-import org.mskcc.cmo.ks.redcap.models.ProjectInfoResponse;
 import org.mskcc.cmo.ks.redcap.models.RedcapAttributeMetadata;
 import org.mskcc.cmo.ks.redcap.models.RedcapProjectAttribute;
-import org.mskcc.cmo.ks.redcap.models.RedcapToken;
 import org.mskcc.cmo.ks.redcap.source.ClinicalDataSource;
-import org.springframework.beans.factory.annotation.Value;
+import org.springframework.beans.factory.annotation.*;
 import org.springframework.http.*;
 import org.springframework.stereotype.Repository;
 import org.springframework.util.LinkedMultiValueMap;
@@ -62,31 +57,12 @@ import org.springframework.web.client.RestTemplate;
 @Repository
 public class ClinicalDataSourceRedcapImpl implements ClinicalDataSource {
 
-    //Do not use these directly .. instead call redcapBaseURI() and redcapApiURI() methods
-    //TODO: move these into a URI management class/bean (singleton)
-    private static URI redcapBaseURI = null;
-    private static URI redcapApiURI = null;
+    @Autowired
+    private RedcapSessionManager redcapSessionManager;
 
-    private static int FILTER_TOKEN_BY_PROJECT_ID  = 0;
-    private static int FILTER_TOKENS_BY_STABLE_ID = 1;
-
-    @Value("${redcap_base_url}")
-    private String redcapBaseUrl;
-    @Value("${redcap_erase_project_data_url_path}")
-    private String redcapEraseProjectDataUrlPath;
-    @Value("${redcap_username}")
-    private String redcapUsername;
-    @Value("${redcap_password}")
-    private String redcapPassword;
-    @Value("${redcap_login_hidden_input_name}")
-    private String redcapLoginHiddenInputName;
-    @Value("${mapping_token}")
-    private String mappingToken;
-    @Value("${metadata_project}")
-    private String metadataProject;
-
-    private Map<String, String> clinicalDataTokens = null; // null until call to fillTokens()
-    private Map<String, String>  clinicalTimelineTokens = null; // null until call to fillTokens()
+    private Map<String, String> clinicalDataTokens = null;
+    private Map<String, String> clinicalTimelineTokens = null;
+    private String metadataToken = null;
     private List<Map<String, String>> records;
     private List<Map<String, String>> timelineRecords;
     List<RedcapAttributeMetadata> metadata;
@@ -98,53 +74,52 @@ public class ClinicalDataSourceRedcapImpl implements ClinicalDataSource {
     private Map<String, List<String>> fullSampleHeader = new HashMap<>();
     String nextClinicalId;
     String nextTimelineId;
-    String metadataToken;
 
     private final Logger log = Logger.getLogger(ClinicalDataSourceRedcapImpl.class);
 
     @Override
     public boolean projectExists(String projectTitle) {
-        checkTokens(projectTitle, FILTER_TOKEN_BY_PROJECT_ID);
+        checkTokensByProjectTitle(projectTitle);
         return (!clinicalDataTokens.isEmpty() || !clinicalTimelineTokens.isEmpty());
     }
 
     @Override
     public List<String> getSampleHeader(String stableId) {
-        checkTokens(stableId, FILTER_TOKENS_BY_STABLE_ID);
+        checkTokensByStableId(stableId);
         getClinicalHeaderData();
         return sampleHeader;
     }
 
     @Override
     public List<String> getPatientHeader(String stableId) {
-        checkTokens(stableId, FILTER_TOKENS_BY_STABLE_ID);
+        checkTokensByStableId(stableId);
         getClinicalHeaderData();
         return patientHeader;
     }
 
     @Override
     public List<String> getTimelineHeader(String stableId) {
-        checkTokens(stableId, FILTER_TOKENS_BY_STABLE_ID);
+        checkTokensByStableId(stableId);
         getTimelineHeaderData();
         return combinedHeader;
     }
 
     @Override
     public List<Map<String, String>> getClinicalData(String stableId) {
-        checkTokens(stableId, FILTER_TOKENS_BY_STABLE_ID);
+        checkTokensByStableId(stableId);
         return records = getClinicalData(false);
     }
 
     @Override
     public List<Map<String, String>> getTimelineData(String stableId) {
-        checkTokens(stableId, FILTER_TOKENS_BY_STABLE_ID);
+        checkTokensByStableId(stableId);
         return timelineRecords = getClinicalData(true);
 
     }
 
     @Override
     public String getNextClinicalProjectTitle(String stableId) {
-        checkTokens(stableId, FILTER_TOKENS_BY_STABLE_ID);
+        checkTokensByStableId(stableId);
         List<String> keys = new ArrayList(clinicalDataTokens.keySet());
         nextClinicalId = keys.get(0);
         return nextClinicalId;
@@ -152,7 +127,7 @@ public class ClinicalDataSourceRedcapImpl implements ClinicalDataSource {
 
     @Override
     public String getNextTimelineProjectTitle(String stableId) {
-        checkTokens(stableId, FILTER_TOKENS_BY_STABLE_ID);
+        checkTokensByStableId(stableId);
         List<String> keys = new ArrayList(clinicalTimelineTokens.keySet());
         nextTimelineId = keys.get(0);
         return nextTimelineId;
@@ -160,19 +135,19 @@ public class ClinicalDataSourceRedcapImpl implements ClinicalDataSource {
 
     @Override
     public boolean hasMoreTimelineData(String stableId) {
-        checkTokens(stableId, FILTER_TOKENS_BY_STABLE_ID);
+        checkTokensByStableId(stableId);
         return !clinicalTimelineTokens.isEmpty();
     }
 
     @Override
     public boolean hasMoreClinicalData(String stableId) {
-        checkTokens(stableId, FILTER_TOKENS_BY_STABLE_ID);
+        checkTokensByStableId(stableId);
         return !clinicalDataTokens.isEmpty();
     }
 
     @Override
     public void importClinicalDataFile(String projectTitle, String filename) {
-        checkTokens(projectTitle, FILTER_TOKEN_BY_PROJECT_ID);
+        checkTokensByProjectTitle(projectTitle);
         String token = clinicalDataTokens.getOrDefault(projectTitle, null);
         if (token == null) {
             token = clinicalTimelineTokens.getOrDefault(projectTitle, null);
@@ -194,41 +169,16 @@ public class ClinicalDataSourceRedcapImpl implements ClinicalDataSource {
                 return;
             }
             String dataForImport = String.join("\n",dataFileContentsCSV.toArray(new String[0])) + "\n";
-            deleteRedcapProjectData(token);
+            redcapSessionManager.deleteRedcapProjectData(token);
             if (dataFileContentsCSV.size() == 1) {
                 log.warn("file " + filename + " contained a single line (presumed to be the header). RedCap project has been cleared (now has 0 records).");
             } else {
-                importClinicalData(token, dataForImport);
+                redcapSessionManager.importClinicalData(token, dataForImport);
                 log.info("import completed, " + Integer.toString(dataFileContentsCSV.size() - 1) + " records imported");
             }
         } catch (IOException e) {
             log.error("IOException thrown while attempting to read file " + filename + " : " + e.getMessage());
         }
-    }
-
-    private URI getRedcapURI() {
-        if (redcapBaseURI == null) {
-            try {
-                redcapBaseURI = new URI(redcapBaseUrl + "/");
-             } catch (URISyntaxException e) {
-                log.error(e.getMessage());
-                throw new RuntimeException(e);
-            }
-        }
-        return redcapBaseURI;
-    }
-
-    private URI getRedcapApiURI() {
-        if (redcapApiURI == null) {
-            URI base = getRedcapURI();
-            redcapApiURI = base.resolve("api/");
-        }
-        return redcapApiURI;
-    }
-
-    private URI getRedcapEraseProjectDataURI(String projectId) {
-        URI base = getRedcapURI();
-        return base.resolve(redcapEraseProjectDataUrlPath + "/?pid=" + projectId + "&action=erase_data");
     }
 
     private void getClinicalHeaderData() {
@@ -272,7 +222,7 @@ public class ClinicalDataSourceRedcapImpl implements ClinicalDataSource {
     }
 
     private List<RedcapAttributeMetadata> getMetadata() {
-        if (metadata  != null) {
+        if (metadata != null) {
             return metadata;
         }
         RestTemplate restTemplate = new RestTemplate();
@@ -285,8 +235,8 @@ public class ClinicalDataSourceRedcapImpl implements ClinicalDataSource {
         uriVariables.add("format", "json");
         uriVariables.add("type", "flat");
 
-        HttpEntity<LinkedMultiValueMap<String, Object>> requestEntity = getRequestEntity(uriVariables);
-        ResponseEntity<RedcapAttributeMetadata[]> responseEntity = restTemplate.exchange(getRedcapApiURI(), HttpMethod.POST, requestEntity, RedcapAttributeMetadata[].class);
+        HttpEntity<LinkedMultiValueMap<String, Object>> requestEntity = redcapSessionManager.getRequestEntity(uriVariables);
+        ResponseEntity<RedcapAttributeMetadata[]> responseEntity = restTemplate.exchange(redcapSessionManager.getRedcapApiURI(), HttpMethod.POST, requestEntity, RedcapAttributeMetadata[].class);
         return Arrays.asList(responseEntity.getBody());
     }
 
@@ -306,13 +256,13 @@ public class ClinicalDataSourceRedcapImpl implements ClinicalDataSource {
         uriVariables.add("type", "flat");
 
         RestTemplate restTemplate = new RestTemplate();
-        HttpEntity<LinkedMultiValueMap<String, Object>> requestEntity = getRequestEntity(uriVariables);
+        HttpEntity<LinkedMultiValueMap<String, Object>> requestEntity = redcapSessionManager.getRequestEntity(uriVariables);
         log.info("Getting attributes for project...");
-        ResponseEntity<RedcapProjectAttribute[]> responseEntity = restTemplate.exchange(getRedcapApiURI(), HttpMethod.POST, requestEntity, RedcapProjectAttribute[].class);
+        ResponseEntity<RedcapProjectAttribute[]> responseEntity = restTemplate.exchange(redcapSessionManager.getRedcapApiURI(), HttpMethod.POST, requestEntity, RedcapProjectAttribute[].class);
         return Arrays.asList(responseEntity.getBody());
     }
 
-    private  List<Map<String, String>> getClinicalData(boolean timelineData) {
+    private List<Map<String, String>> getClinicalData(boolean timelineData) {
         String projectToken;
         if(timelineData) {
             projectToken = clinicalTimelineTokens.remove(nextTimelineId);
@@ -329,9 +279,9 @@ public class ClinicalDataSourceRedcapImpl implements ClinicalDataSource {
         uriVariables.add("type", "flat");
 
         RestTemplate restTemplate = new RestTemplate();
-        HttpEntity<LinkedMultiValueMap<String, Object>> requestEntity = getRequestEntity(uriVariables);
+        HttpEntity<LinkedMultiValueMap<String, Object>> requestEntity = redcapSessionManager.getRequestEntity(uriVariables);
         log.info("Getting data for project...");
-        ResponseEntity<ObjectNode[]> responseEntity = restTemplate.exchange(getRedcapApiURI(), HttpMethod.POST, requestEntity, ObjectNode[].class);
+        ResponseEntity<ObjectNode[]> responseEntity = restTemplate.exchange(redcapSessionManager.getRedcapApiURI(), HttpMethod.POST, requestEntity, ObjectNode[].class);
         List<Map<String, String>> responses = new ArrayList<>();
 
         for(ObjectNode response : responseEntity.getBody())
@@ -356,55 +306,23 @@ public class ClinicalDataSourceRedcapImpl implements ClinicalDataSource {
         return header;
     }
 
-    private void fillTokens(String tokenFilterString, int tokenFilterMode) {
-        RestTemplate restTemplate = new RestTemplate();
-        clinicalTimelineTokens = new HashMap<>();
-        clinicalDataTokens = new HashMap<>();
+    private boolean gotTokens() {
+        return clinicalTimelineTokens != null && clinicalDataTokens != null && metadataToken != null;
+    }
 
-        log.info("Getting tokens for clinical data processor...");
-
-        LinkedMultiValueMap<String, String> uriVariables = new LinkedMultiValueMap<>();
-        uriVariables.add("token", mappingToken);
-        uriVariables.add("content", "record");
-        uriVariables.add("format", "json");
-        uriVariables.add("type", "flat");
-
-        HttpEntity<LinkedMultiValueMap<String, Object>> requestEntity = getRequestEntity(uriVariables);
-        ResponseEntity<RedcapToken[]> responseEntity = restTemplate.exchange(getRedcapApiURI(), HttpMethod.POST, requestEntity, RedcapToken[].class);
-
-        for (RedcapToken token : responseEntity.getBody()) {
-            if (token.getStableId().equals(metadataProject)) {
-                metadataToken = token.getApiToken();
-                continue;
-            }
-            if ((tokenFilterMode == FILTER_TOKENS_BY_STABLE_ID && token.getStableId().equals(tokenFilterString)) ||
-                    (tokenFilterMode == FILTER_TOKEN_BY_PROJECT_ID  && token.getStudyId().equals(tokenFilterString))) {
-                if (token.getStudyId().toUpperCase().contains("TIMELINE")) {
-                    clinicalTimelineTokens.put(token.getStudyId(), token.getApiToken());
-                }
-                else {
-                    clinicalDataTokens.put(token.getStudyId(), token.getApiToken());
-                }
-            }
+    private void checkTokensByProjectTitle(String projectTitle) {
+        if (!gotTokens()) {
+            clinicalTimelineTokens = redcapSessionManager.getTimelineTokenMapByProjectTitle(projectTitle);
+            clinicalDataTokens = redcapSessionManager.getClinicalTokenMapByProjectTitle(projectTitle);
+            metadataToken = redcapSessionManager.getMetadataTokenByProjectTitle(projectTitle);
         }
     }
 
-    private HttpEntity getRequestEntity(LinkedMultiValueMap<String, String> uriVariables)
-    {
-        HttpHeaders headers = new HttpHeaders();
-        headers.setContentType(MediaType.APPLICATION_FORM_URLENCODED);
-        headers.setAccept(Arrays.asList(MediaType.APPLICATION_JSON));
-        return new HttpEntity<LinkedMultiValueMap<String, String>>(uriVariables, headers);
-    }
-
-    private boolean gotTokens() {
-        // token maps are null until created during call to fillTokens()
-        return clinicalTimelineTokens != null || clinicalDataTokens != null;
-    }
-
-    private void checkTokens(String tokenFilterString, int tokenFilterMode) {
+    private void checkTokensByStableId(String stableId) {
         if (!gotTokens()) {
-            fillTokens(tokenFilterString, tokenFilterMode);
+            clinicalTimelineTokens = redcapSessionManager.getTimelineTokenMapByStableId(stableId);
+            clinicalDataTokens = redcapSessionManager.getClinicalTokenMapByStableId(stableId);
+            metadataToken = redcapSessionManager.getMetadataTokenByStableId(stableId);
         }
     }
 
@@ -474,72 +392,6 @@ public class ClinicalDataSourceRedcapImpl implements ClinicalDataSource {
         return fullSampleHeader;
     }
 
-    private String getSessionCookieFromRedcap() {
-        RestTemplate restTemplate = new RestTemplate();
-        LinkedMultiValueMap<String, String> loginUriVariables = new LinkedMultiValueMap<>();
-        loginUriVariables.add("username", redcapUsername);
-        loginUriVariables.add("password", redcapPassword);
-        loginUriVariables.add(redcapLoginHiddenInputName, "");
-        loginUriVariables.add("submitted", "1");
-        HttpEntity<LinkedMultiValueMap<String, Object>> loginRequestEntity = getRequestEntity(loginUriVariables);
-        ResponseEntity<String> loginResponseEntity = restTemplate.exchange(getRedcapURI(), HttpMethod.POST, loginRequestEntity, String.class);
-        HttpStatus responseStatus = loginResponseEntity.getStatusCode();
-        if (!responseStatus.is2xxSuccessful() && !responseStatus.is3xxRedirection()) {
-            log.warn("RedCap login with username/password/hiddenInput attempt failed. HTTP status code = " + Integer.toString(loginResponseEntity.getStatusCode().value()));
-            return null;
-        }
-        Predicate<String> notDeleted = (String s) -> !s.contains("deleted");
-        List<String> cookies = loginResponseEntity.getHeaders().get("Set-Cookie").stream().filter(notDeleted).collect(Collectors.toList());
-        if (cookies.size() < 1) {
-            log.warn("RedCap login succeeded but no Set-Cookie header field was included.");
-            return null;
-        }
-        return cookies.get(0).split(";")[0];
-    }
-
-    private String getProjectIdFromRedcap(String token, String cookie) {
-        RestTemplate restTemplate = new RestTemplate();
-        LinkedMultiValueMap<String, String> projectInfoUriVariables = new LinkedMultiValueMap<>();
-        projectInfoUriVariables.add("token", token);
-        projectInfoUriVariables.add("content", "project");
-        projectInfoUriVariables.add("format", "json");
-        projectInfoUriVariables.add("returnFormat", "json");
-        HttpEntity<LinkedMultiValueMap<String, Object>> projectInfoRequestEntity = getRequestEntity(projectInfoUriVariables);
-        HttpHeaders headers = new HttpHeaders();
-        headers.add(HttpHeaders.COOKIE, cookie);
-        HttpEntity<?> rq = new HttpEntity<>(headers);
-        ResponseEntity<ProjectInfoResponse> projectInfoResponseEntity = restTemplate.exchange(getRedcapApiURI(), HttpMethod.POST, projectInfoRequestEntity, ProjectInfoResponse.class);
-        HttpStatus responseStatus = projectInfoResponseEntity.getStatusCode();
-        if (!responseStatus.is2xxSuccessful() && !responseStatus.is3xxRedirection()) {
-            log.warn("RedCap request for project data failed. HTTP status code = " + Integer.toString(projectInfoResponseEntity.getStatusCode().value()));
-            return null;
-        }
-        return projectInfoResponseEntity.getBody().getProjectId();
-    }
-
-    private void deleteRedcapProjectData(String cookie, String projectId) {
-        RestTemplate restTemplate = new RestTemplate();
-        HttpHeaders headers = new HttpHeaders();
-        headers.add(HttpHeaders.COOKIE, cookie);
-        HttpEntity<?> rq = new HttpEntity<>(headers);
-        HttpEntity<?> response = restTemplate.exchange(getRedcapEraseProjectDataURI(projectId), HttpMethod.GET, rq, String.class);
-    }
-
-    private void deleteRedcapProjectData(String token) {
-        String cookie = getSessionCookieFromRedcap();
-        if (cookie == null) {
-            log.warn("RedCap session cookie not available; unable to delete project data");
-            throw new RuntimeException("RedCap session cookie not available; unable to delete project data");
-        }
-        String projectId = getProjectIdFromRedcap(token, cookie);
-        if (projectId == null) {
-            log.warn("ProjectId not available from RedCap getProjectData API request");
-            throw new RuntimeException("ProjectId not available from RedCap getProjectData API request");
-        }
-        log.info("deleting all records for RedCap projectId: " + projectId);
-        deleteRedcapProjectData(cookie, projectId);
-    }
-
     private List<String> readClinicalFile(File file) throws IOException {
         LinkedList<String> lineList = new LinkedList<>();
         BufferedReader bufferedReader = new BufferedReader(new FileReader(file));
@@ -568,25 +420,6 @@ public class ClinicalDataSourceRedcapImpl implements ClinicalDataSource {
             csvLines.add(String.join(",", csvFields));
         }
         return csvLines;
-    }
-
-    public void importClinicalData(String token, String dataForImport) {
-        log.info("importing data ... (" + dataForImport.length() + " characters)");
-        RestTemplate restTemplate = new RestTemplate();
-        LinkedMultiValueMap<String, String> importRecordUriVariables = new LinkedMultiValueMap<>();
-        importRecordUriVariables.add("token", token);
-        importRecordUriVariables.add("content", "record");
-        importRecordUriVariables.add("format", "csv");
-        importRecordUriVariables.add("overwriteBehavior", "overwrite");
-        importRecordUriVariables.add("data", dataForImport);
-        HttpEntity<LinkedMultiValueMap<String, Object>> importRecordRequestEntity = getRequestEntity(importRecordUriVariables);
-        ResponseEntity<String> importRecordResponseEntity = restTemplate.exchange(getRedcapApiURI(), HttpMethod.POST, importRecordRequestEntity, String.class);
-        HttpStatus responseStatus = importRecordResponseEntity.getStatusCode();
-        if (!responseStatus.is2xxSuccessful() && !responseStatus.is3xxRedirection()) {
-            log.warn("RedCap import record API call failed. HTTP status code = " + Integer.toString(importRecordResponseEntity.getStatusCode().value()));
-            throw new RuntimeException("RedCap import record API call failed. HTTP status code");
-        }
-        log.info("Return from call to Import Recap Record API: " + importRecordResponseEntity.getBody());
     }
 
     public static void main(String[] args) {}
