+++ redcap/redcap_source/src/main/java/org/mskcc/cmo/ks/redcap/source/internal/RedcapSessionManager.java
@@ -1,0 +1,332 @@
+/*
+ * Copyright (c) 2017 Memorial Sloan-Kettering Cancer Center.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY, WITHOUT EVEN THE IMPLIED WARRANTY OF MERCHANTABILITY OR FITNESS
+ * FOR A PARTICULAR PURPOSE. The software and documentation provided hereunder
+ * is on an "as is" basis, and Memorial Sloan-Kettering Cancer Center has no
+ * obligations to provide maintenance, support, updates, enhancements or
+ * modifications. In no event shall Memorial Sloan-Kettering Cancer Center be
+ * liable to any party for direct, indirect, special, incidental or
+ * consequential damages, including lost profits, arising out of the use of this
+ * software and its documentation, even if Memorial Sloan-Kettering Cancer
+ * Center has been advised of the possibility of such damage.
+ */
+
+/*
+ * This file is part of cBioPortal CMO-Pipelines.
+ *
+ * cBioPortal is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+*/
+
+package org.mskcc.cmo.ks.redcap.source.internal;
+
+import java.io.*;
+import java.net.*;
+import java.util.*;
+import java.util.function.Predicate;
+import java.util.stream.Collectors;
+import org.apache.log4j.Logger;
+import org.mskcc.cmo.ks.redcap.models.ProjectInfoResponse;
+import org.mskcc.cmo.ks.redcap.models.RedcapToken;
+import org.mskcc.cmo.ks.redcap.source.ClinicalDataSource;
+import org.springframework.beans.factory.annotation.*;
+import org.springframework.http.*;
+import org.springframework.stereotype.Repository;
+import org.springframework.util.LinkedMultiValueMap;
+import org.springframework.web.client.RestTemplate;
+
+@Repository
+public class RedcapSessionManager {
+
+    private static URI redcapBaseURI = null;
+    private static URI redcapApiURI = null;
+
+    private static int FILTER_TOKENS_METHOD_UNINITIALIZED = -1;
+    private static int FILTER_TOKENS_BY_PROJECT_TITLE = 0;
+    private static int FILTER_TOKENS_BY_STABLE_ID = 1;
+    private static int FILTER_TOKENS_NO_CANCER_STUDIES = 2;
+
+    @Value("${redcap_base_url}")
+    private String redcapBaseUrl;
+    @Value("${redcap_erase_project_data_url_path}")
+    private String redcapEraseProjectDataUrlPath;
+    @Value("${redcap_username}")
+    private String redcapUsername;
+    @Value("${redcap_password}")
+    private String redcapPassword;
+    @Value("${redcap_login_hidden_input_name}")
+    private String redcapLoginHiddenInputName;
+    @Value("${mapping_token}")
+    private String mappingToken;
+    @Value("${metadata_project}")
+    private String metadataProjectStableId;
+    @Value("${namespace_project}")
+    private String namespaceProjectStableId;
+
+    private Map<String, String> clinicalDataTokens = null;
+    private Map<String, String> clinicalTimelineTokens = null;
+    String metadataToken = null;
+    String namespaceToken = null;
+    int selectedTokenFilterMode = FILTER_TOKENS_METHOD_UNINITIALIZED;
+    String selectedTokenFilterString = null;
+
+    private final Logger log = Logger.getLogger(RedcapSessionManager.class);
+
+    // SECTION : URI construction
+
+    public URI getRedcapURI() {
+        if (redcapBaseURI == null) {
+            try {
+                redcapBaseURI = new URI(redcapBaseUrl + "/");
+             } catch (URISyntaxException e) {
+                log.error(e.getMessage());
+                throw new RuntimeException(e);
+            }
+        }
+        return redcapBaseURI;
+    }
+
+    public URI getRedcapApiURI() {
+        if (redcapApiURI == null) {
+            URI base = getRedcapURI();
+            redcapApiURI = base.resolve("api/");
+        }
+        return redcapApiURI;
+    }
+
+    private URI getRedcapEraseProjectDataURI(String projectId) {
+        URI base = getRedcapURI();
+        return base.resolve(redcapEraseProjectDataUrlPath + "/?pid=" + projectId + "&action=erase_data");
+    }
+
+    // SECTION : RedCap project token handling
+
+    public Map<String, String> getClinicalTokenMapByProjectTitle(String projectTitle) {
+        checkTokens(projectTitle, FILTER_TOKENS_BY_PROJECT_TITLE);
+        return new HashMap<String, String>(clinicalDataTokens);
+    }
+
+    public Map<String, String> getTimelineTokenMapByProjectTitle(String projectTitle) {
+        checkTokens(projectTitle, FILTER_TOKENS_BY_PROJECT_TITLE);
+        return new HashMap<String, String>(clinicalTimelineTokens);
+    }
+
+    public String getMetadataTokenByProjectTitle(String projectTitle) {
+        checkTokens(projectTitle, FILTER_TOKENS_BY_PROJECT_TITLE);
+        return metadataToken;
+    }
+
+    public String getNamespaceTokenByProjectTitle(String projectTitle) {
+        checkTokens(projectTitle, FILTER_TOKENS_BY_PROJECT_TITLE);
+        return namespaceToken;
+    }
+
+    public Map<String, String> getClinicalTokenMapByStableId(String stableId) {
+        checkTokens(stableId, FILTER_TOKENS_BY_STABLE_ID);
+        return new HashMap<String, String>(clinicalDataTokens);
+    }
+
+    public Map<String, String> getTimelineTokenMapByStableId(String stableId) {
+        checkTokens(stableId, FILTER_TOKENS_BY_STABLE_ID);
+        return new HashMap<String, String>(clinicalTimelineTokens);
+    }
+
+    public String getMetadataTokenByStableId(String stableId) {
+        checkTokens(stableId, FILTER_TOKENS_BY_STABLE_ID);
+        return metadataToken;
+    }
+
+    public String getNamespaceTokenByStableId(String stableId) {
+        checkTokens(stableId, FILTER_TOKENS_BY_STABLE_ID);
+        return namespaceToken;
+    }
+
+    public String getMetadataTokenFilterAllStudiesIfNecessary() {
+        if (metadataToken != null) {
+            return metadataToken;
+        }
+        checkTokens("", FILTER_TOKENS_NO_CANCER_STUDIES);
+        return metadataToken;
+    }
+
+    public String getNamespaceTokenFilterAllStudiesIfNecessary() {
+        if (namespaceToken != null) {
+            return namespaceToken;
+        }
+        checkTokens("", FILTER_TOKENS_NO_CANCER_STUDIES);
+        return namespaceToken;
+    }
+
+    public boolean tokensAreInitialized() {
+        return clinicalTimelineTokens != null && clinicalDataTokens != null && metadataToken != null;
+    }
+
+    private void checkTokens(String tokenFilterString, int tokenFilterMode) {
+        if (selectedTokenFilterMode == FILTER_TOKENS_METHOD_UNINITIALIZED) {
+            selectedTokenFilterMode = tokenFilterMode;
+            selectedTokenFilterString = tokenFilterString;
+        }
+        if (selectedTokenFilterMode != tokenFilterMode) {
+            throw new RuntimeException("Error : RedCap tokenFilterMode has changed from " +
+                    Integer.toString(selectedTokenFilterMode) + " to " +
+                    Integer.toString(tokenFilterMode) +
+                    " : once tokens have been fetched and filtered, the filtering mode cannot be changed within the same instance of RedcapSessionManager");
+        }
+        if (selectedTokenFilterString != tokenFilterString) {
+            throw new RuntimeException("Error : RedCap tokenFilterMode has changed from " +
+                    selectedTokenFilterString + " to " + tokenFilterString +
+                    Integer.toString(tokenFilterMode) +
+                    " : once tokens have been fetched and filtered, the filtering mode cannot be changed within the same instance of RedcapSessionManager");
+        }
+        if (!tokensAreInitialized()) {
+            fillTokens();
+        }
+    }
+
+    private void fillTokens() {
+        RestTemplate restTemplate = new RestTemplate();
+        clinicalTimelineTokens = new HashMap<>();
+        clinicalDataTokens = new HashMap<>();
+
+        log.info("Getting tokens for clinical data processor...");
+
+        LinkedMultiValueMap<String, String> uriVariables = new LinkedMultiValueMap<>();
+        uriVariables.add("token", mappingToken);
+        uriVariables.add("content", "record");
+        uriVariables.add("format", "json");
+        uriVariables.add("type", "flat");
+
+        HttpEntity<LinkedMultiValueMap<String, Object>> requestEntity = getRequestEntity(uriVariables);
+        ResponseEntity<RedcapToken[]> responseEntity = restTemplate.exchange(getRedcapApiURI(), HttpMethod.POST, requestEntity, RedcapToken[].class);
+
+        for (RedcapToken token : responseEntity.getBody()) {
+            if (token.getStableId().equals(metadataProjectStableId)) {
+                metadataToken = token.getApiToken();
+                continue;
+            }
+            if (token.getStableId().equals(namespaceProjectStableId)) {
+                namespaceToken = token.getApiToken();
+                continue;
+            }
+            if ((selectedTokenFilterMode == FILTER_TOKENS_BY_STABLE_ID && token.getStableId().equals(selectedTokenFilterString)) ||
+                    (selectedTokenFilterMode == FILTER_TOKENS_BY_PROJECT_TITLE  && token.getStudyId().equals(selectedTokenFilterString))) {
+                if (token.getStudyId().toUpperCase().contains("TIMELINE")) {
+                    clinicalTimelineTokens.put(token.getStudyId(), token.getApiToken());
+                }
+                else {
+                    clinicalDataTokens.put(token.getStudyId(), token.getApiToken());
+                }
+            }
+        }
+        if (metadataToken == null) {
+            throw new RuntimeException("Error : could not locate token for RedCap metadata project with stable id \"" + metadataProjectStableId + "\"");
+        }
+    }
+
+    // SECTION : utility functions for doing RedCap specific requests
+
+    public void deleteRedcapProjectData(String token) {
+        String cookie = getSessionCookieFromRedcap();
+        if (cookie == null) {
+            log.warn("RedCap session cookie not available; unable to delete project data");
+            throw new RuntimeException("RedCap session cookie not available; unable to delete project data");
+        }
+        String projectId = getProjectIdFromRedcap(token, cookie);
+        if (projectId == null) {
+            log.warn("ProjectId not available from RedCap getProjectData API request");
+            throw new RuntimeException("ProjectId not available from RedCap getProjectData API request");
+        }
+        log.info("deleting all records for RedCap projectId: " + projectId);
+        deleteRedcapProjectData(cookie, projectId);
+    }
+
+    public void importClinicalData(String token, String dataForImport) {
+        log.info("importing data ... (" + dataForImport.length() + " characters)");
+        RestTemplate restTemplate = new RestTemplate();
+        LinkedMultiValueMap<String, String> importRecordUriVariables = new LinkedMultiValueMap<>();
+        importRecordUriVariables.add("token", token);
+        importRecordUriVariables.add("content", "record");
+        importRecordUriVariables.add("format", "csv");
+        importRecordUriVariables.add("overwriteBehavior", "overwrite");
+        importRecordUriVariables.add("data", dataForImport);
+        HttpEntity<LinkedMultiValueMap<String, Object>> importRecordRequestEntity = getRequestEntity(importRecordUriVariables);
+        ResponseEntity<String> importRecordResponseEntity = restTemplate.exchange(getRedcapApiURI(), HttpMethod.POST, importRecordRequestEntity, String.class);
+        HttpStatus responseStatus = importRecordResponseEntity.getStatusCode();
+        if (!responseStatus.is2xxSuccessful() && !responseStatus.is3xxRedirection()) {
+            log.warn("RedCap import record API call failed. HTTP status code = " + Integer.toString(importRecordResponseEntity.getStatusCode().value()));
+            throw new RuntimeException("RedCap import record API call failed. HTTP status code");
+        }
+        log.info("Return from call to Import Recap Record API: " + importRecordResponseEntity.getBody());
+    }
+
+    public HttpEntity getRequestEntity(LinkedMultiValueMap<String, String> uriVariables) {
+        HttpHeaders headers = new HttpHeaders();
+        headers.setContentType(MediaType.APPLICATION_FORM_URLENCODED);
+        headers.setAccept(Arrays.asList(MediaType.APPLICATION_JSON));
+        return new HttpEntity<LinkedMultiValueMap<String, String>>(uriVariables, headers);
+    }
+
+    private String getSessionCookieFromRedcap() {
+        RestTemplate restTemplate = new RestTemplate();
+        LinkedMultiValueMap<String, String> loginUriVariables = new LinkedMultiValueMap<>();
+        loginUriVariables.add("username", redcapUsername);
+        loginUriVariables.add("password", redcapPassword);
+        loginUriVariables.add(redcapLoginHiddenInputName, "");
+        loginUriVariables.add("submitted", "1");
+        HttpEntity<LinkedMultiValueMap<String, Object>> loginRequestEntity = getRequestEntity(loginUriVariables);
+        ResponseEntity<String> loginResponseEntity = restTemplate.exchange(getRedcapURI(), HttpMethod.POST, loginRequestEntity, String.class);
+        HttpStatus responseStatus = loginResponseEntity.getStatusCode();
+        if (!responseStatus.is2xxSuccessful() && !responseStatus.is3xxRedirection()) {
+            log.warn("RedCap login with username/password/hiddenInput attempt failed. HTTP status code = " + Integer.toString(loginResponseEntity.getStatusCode().value()));
+            return null;
+        }
+        Predicate<String> notDeleted = (String s) -> !s.contains("deleted");
+        List<String> cookies = loginResponseEntity.getHeaders().get("Set-Cookie").stream().filter(notDeleted).collect(Collectors.toList());
+        if (cookies.size() < 1) {
+            log.warn("RedCap login succeeded but no Set-Cookie header field was included.");
+            return null;
+        }
+        return cookies.get(0).split(";")[0];
+    }
+
+    private String getProjectIdFromRedcap(String token, String cookie) {
+        RestTemplate restTemplate = new RestTemplate();
+        LinkedMultiValueMap<String, String> projectInfoUriVariables = new LinkedMultiValueMap<>();
+        projectInfoUriVariables.add("token", token);
+        projectInfoUriVariables.add("content", "project");
+        projectInfoUriVariables.add("format", "json");
+        projectInfoUriVariables.add("returnFormat", "json");
+        HttpEntity<LinkedMultiValueMap<String, Object>> projectInfoRequestEntity = getRequestEntity(projectInfoUriVariables);
+        HttpHeaders headers = new HttpHeaders();
+        headers.add(HttpHeaders.COOKIE, cookie);
+        HttpEntity<?> rq = new HttpEntity<>(headers);
+        ResponseEntity<ProjectInfoResponse> projectInfoResponseEntity = restTemplate.exchange(getRedcapApiURI(), HttpMethod.POST, projectInfoRequestEntity, ProjectInfoResponse.class);
+        HttpStatus responseStatus = projectInfoResponseEntity.getStatusCode();
+        if (!responseStatus.is2xxSuccessful() && !responseStatus.is3xxRedirection()) {
+            log.warn("RedCap request for project data failed. HTTP status code = " + Integer.toString(projectInfoResponseEntity.getStatusCode().value()));
+            return null;
+        }
+        return projectInfoResponseEntity.getBody().getProjectId();
+    }
+
+    private void deleteRedcapProjectData(String cookie, String projectId) {
+        RestTemplate restTemplate = new RestTemplate();
+        HttpHeaders headers = new HttpHeaders();
+        headers.add(HttpHeaders.COOKIE, cookie);
+        HttpEntity<?> rq = new HttpEntity<>(headers);
+        HttpEntity<?> response = restTemplate.exchange(getRedcapEraseProjectDataURI(projectId), HttpMethod.GET, rq, String.class);
+    }
+
+}
